# lib/api-zod/src/generated/types/localLoginInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LocalLoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}